requirements.txt
autogen==0.8.1
arxiv==2.1.3
requests==2.32.3
pdfminer.six==20240706
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
schedule==1.2.2
openai==1.66.3
sqlalchemy==2.0.39
